ai-agent-lite/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
prometheus-client>=0.20.0
pydantic>=2.0
pyyaml>=6.0